pyproject.toml
[tool.poetry]
name = "django-strawberry-cookbook"
version = "0.1.0"
description = "Common patterns when creating a Django Strawberry (GraphQL) API"
authors = ["Dylan Stein <dylan@steins.studio>"]
license = "BSD 3-Clause License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "5.0b1"
strawberry-graphql = {extras = ["cli"], version = "^0.211.1"}
uvicorn = "^0.23.2"
strawberry-graphql-django = "^0.22.0"
whitenoise = "^6.6.0"
django-cors-headers = "^4.3.0"
django-extensions = "^3.2.3"
django-environ = "^0.11.2"
psycopg = "^3.1.12"
ipdb = "^0.13.13"
django-guardian = "^2.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
ipython = "^8.17.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
